refactor!: align domain types with what the API returns
- remove Zone.SOASerial and Zone.DNSSECSigned: no Poweradmin release
  returns soa_serial or dnssec_signed, so both were always zero; the
  DNSSEC status is available via ZoneClient.GetDNSSEC
- add Zone.CreatedAt, which every zone endpoint returns
- document that list results only carry ID, Name, Type and CreatedAt,
  and drop the outdated note that /v2/zones is the only wrapped list
- Record.ZoneID is an int like every other zone ID in the SDK
- add GroupMember.Fullname/Email and GroupZone.CreatedAt, which were
  decoded but dropped during conversion

BREAKING CHANGE: Zone.SOASerial and Zone.DNSSECSigned were removed, and
Record.ZoneID changed from int64 to int.

+++ poweradmin/record.go
@@ -12,7 +12,7 @@ import (
 // Record represents a DNS record in Poweradmin.
 type Record struct {
 	ID       string
-	ZoneID   int64
+	ZoneID   int
 	Name     string
 	Type     string
 	Content  string
